+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Based all gltf data classes on `BaseGLTFDataClass`
 * Fixed `Color.__get___` AttributeError.
+* Fixed  `RhinoSurface.curvature_at` not returning a Vector, but a Rhino SurfaceCurvature class object 
 * Fixed `cylinder_to_rhino` conversion to match `compas.geometry.Cylinder` location.
 * Changed identification of cylinder brep face to non-zero in `compas_rhino.conversions.cylinder.Cylinder`.
 * Changed linter to `black`.
@@ -35,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed bug in `_core.tangent` where the Circle's Plane was accessed instead of the centre.
 * Fixed the `test_tangent` to work with a properly defined circle
 * `RhinoBrep` serialization works now with surface types other than NURBS.
+
 ### Removed